extrait des définitions communes pour les exports
